global_snowflake_utils/requirements.txt
# === Core Snowflake & Database ===
snowflake-connector-python>=3.5.0
snowflake-sqlalchemy>=1.4.0
sqlalchemy>=1.4.0

# === Data Analysis & Processing ===
pandas>=1.5.0
numpy>=1.21.0

# === Configuration & Environment ===
python-dotenv>=1.0.0
cryptography>=3.4.8

# === Data Visualization (Optional but Useful) ===
plotly>=5.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# === Jupyter/Notebook Support (Optional) ===
jupyter>=1.0.0
ipykernel>=6.0.0

# === Utilities ===
requests>=2.28.0
pytz>=2022.1

# === Development & Testing (Optional) ===
pytest>=7.0.0
black>=22.0.0
isort>=5.10.0

# === Performance (Optional) ===
openpyxl>=3.0.0  # For Excel file operations
xlsxwriter>=3.0.0  # For Excel export with formatting
